Using the more performant spew package instead of the pretty package.
The pretty package is great, but it is a bit slower when formatting
the arguments that it is passed. For efficiency reasons, modifying the
code to use the more efficient spew package. To have (roughly) the same
output format - there are some minor differences - we need to use the
Sdump() method of the spew package instead of the Sprint() one.

Also, to avoid going over the object internals twice, we store the
result locally and pass it to the defer function instead.